// src/app/BusinessModel/Tennis/tennis-enum.ts
export enum TennisEnum {
    Date = 'Date',
    ContestGroupId = 'ContestGroupId',
    ContestGroupText = 'ContestGroupText',
    RoundName = 'RoundName',
    RoundId = 'RoundId',
    StatsTypeId = 'StatsTypeId',
    StatsTypeValue = 'StatsTypeValue',
    OrgId = 'OrgId',
    OrgName = 'OrgName',
    SelectedHomeTeamId = 'SelectedHomeTeamId',
    SelectedAwayTeamId = 'selectedAwayTeamId',
    SelectedHomeTeamName = 'SelectedHomeTeamName',
    SelectedAwayTeamName = 'selectedAwayTeamName',
}